+++ R/ard_hierarchical.R
@@ -57,7 +57,7 @@ ard_hierarchical_count <- function(data, ...) {
 ard_hierarchical.data.frame <- function(data,
                                         variables,
                                         by = dplyr::group_vars(data),
-                                        statistic = everything() ~ categorical_summary_fns(),
+                                        statistic = everything() ~ c("n", "N", "p"),
                                         denominator = NULL, fmt_fn = NULL,
                                         stat_label = everything() ~ default_stat_labels(),
                                         id = NULL,
@@ -165,7 +165,7 @@ ard_hierarchical_count.data.frame <- function(data,
           variables = "...ard_dummy_for_counting...",
           by = all_of(by),
           strata = all_of(variables[seq_len(i)]),
-          statistic = everything() ~ categorical_summary_fns("n"),
+          statistic = everything() ~ "n",
           fmt_fn = fmt_fn,
           stat_label = stat_label
         ) |>